chore: migrate rust/inter-canister-calls to icp-cli
