--- rust/inter-canister-calls/.devcontainer/devcontainer.json
@@ -1,20 +1,0 @@
-{
-  "name": "ICP Dev Environment",
-  "image": "ghcr.io/dfinity/icp-dev-env-slim:22",
-  "forwardPorts": [4943, 5173],
-  "portsAttributes": {
-    "4943": {
-      "label": "dfx",
-      "onAutoForward": "ignore"
-    },
-    "5173": {
-      "label": "vite",
-      "onAutoForward": "openBrowser"
-    }
-  },
-  "customizations": {
-    "vscode": {
-      "extensions": ["dfinity-foundation.vscode-motoko"]
-    }
-  }
-}
