+++ rust/inter-canister-calls/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = [ "src/caller", "src/counter" ]
+members = ["caller", "counter"]
 resolver = "2"
