+++ rust/inter-canister-calls/Makefile
@@ -1,37 +1,69 @@
-.PHONY: all
-all: test
-
-.PHONY: deploy
-.SILENT: deploy
-build:
-	dfx deploy
-SHELL := /bin/bash
-.SHELLFLAGS := -euo pipefail -c
 .PHONY: test
-.SILENT: test
-test: build
-	dfx canister call counter set "( 0 : nat )" | grep '()' && echo 'PASS'
-	dfx canister call caller call_get_and_set "( principal \"`dfx canister id counter`\", 42 : nat )" | grep '(0 : nat)' && echo 'PASS'
-	dfx canister call caller set_then_get "( principal \"`dfx canister id counter`\", 7 : nat )" | grep '(7 : nat)' && echo 'PASS'
-	dfx canister call counter get | grep '(7 : nat)' && echo 'PASS'
-	dfx canister call caller call_increment "( principal \"`dfx canister id counter`\" )" | grep '(variant { Ok })' && echo 'PASS'
-	dfx canister call counter get | grep '(8 : nat)' && echo 'PASS'
-	dfx canister call caller call_get "( principal \"`dfx canister id counter`\" )" | grep '(variant { Ok = 8 : nat })' && echo 'PASS'
-	dfx canister call caller stubborn_set "( principal \"`dfx canister id counter`\", 42 : nat )" | grep '(variant { Ok })' && echo 'PASS'
-	dfx canister call caller call_get "( principal \"`dfx canister id counter`\" )" | grep '(variant { Ok = 42 : nat })' && echo 'PASS'
-	# We'll test that the counter's balance increases when calling send_cycles to it.
-	# Storing the original balance in a file is the easiest with make.
-	echo $$(dfx canister status counter | awk '/Balance/ { gsub(/_/, "", $$2); print $$2 }') > .counter_before
-	# Send 10% of the caller's available cycles
-	CYCLES_TO_SEND=$$(dfx canister status caller | awk '/Balance/ { gsub(/_/, "", $$2); printf("%d", int($$2/10)) }'); \
-	echo "Sending $$CYCLES_TO_SEND cycles to counter canister"; \
-	dfx canister call caller send_cycles "( principal \"`dfx canister id counter`\", $$CYCLES_TO_SEND: nat64 )" | grep '(variant { Ok })' && echo PASS
-	# Check that the counter's balance increased
-	AFTER=$$(dfx canister status counter | awk '/Balance/ { gsub(/_/, "", $$2); print $$2 }'); \
-	BEFORE=$$(cat .counter_before); \
-	test $$AFTER -gt $$BEFORE && echo "PASS (cycles increased: $$BEFORE -> $$AFTER)"
 
-.PHONY: clean
-.SILENT: clean
-clean:
-	rm -rf .dfx
+test:
+	@echo "=== Test 1: Reset counter to 0 ==="
+	@result=$$(icp canister call counter set '(0 : nat)') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '()' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 2: call_get_and_set returns old value and sets new ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller call_get_and_set "(principal \"$$COUNTER_ID\", 42 : nat)") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(0 : nat)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 3: set_then_get sets value and returns it ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller set_then_get "(principal \"$$COUNTER_ID\", 7 : nat)") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(7 : nat)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 4: counter value persisted after set_then_get ==="
+	@result=$$(icp canister call --query counter get '()') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(7 : nat)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 5: call_increment increments counter ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller call_increment "(principal \"$$COUNTER_ID\")") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q 'Ok' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 6: counter value is 8 after increment ==="
+	@result=$$(icp canister call --query counter get '()') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(8 : nat)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 7: call_get returns current counter value ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller call_get "(principal \"$$COUNTER_ID\")") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q 'Ok = 8' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 8: stubborn_set retries until success ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller stubborn_set "(principal \"$$COUNTER_ID\", 42 : nat)") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q 'Ok' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 9: counter value is 42 after stubborn_set ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller call_get "(principal \"$$COUNTER_ID\")") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q 'Ok = 42' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 10: send_cycles increases counter canister balance ==="
+	@COUNTER_ID=$$(icp canister id counter) && \
+	result=$$(icp canister call caller send_cycles "(principal \"$$COUNTER_ID\", 1_000_000_000_000 : nat64)") && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q 'Ok' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
