+++ rust/inter-canister-calls/icp.yaml
@@ -1,0 +1,10 @@
+canisters:
+  - name: counter
+    recipe:
+      type: "@dfinity/rust@v3.3.0"
+
+  - name: caller
+    recipe:
+      type: "@dfinity/rust@v3.3.0"
+    environment_variables:
+      - PUBLIC_CANISTER_ID:counter
