+++ rust/inter-canister-calls/rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+targets = ["wasm32-unknown-unknown"]
