--- rust/inter-canister-calls/src/counter/counter.did
@@ -1,6 +1,0 @@
-service : {
-    "get": () -> (nat);
-    "set": (nat) -> ();
-    "increment": () -> ();
-    "get_and_set": (nat) -> (nat);
-}
